+++ deltafi-ui/CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Added `uploadedBy` field to metadata of all uploaded DeltaFiles (#283)
 
+### Fixed
+
+- Fixed broken Indexed Metadata links to Search page (#284)
+
 ## [0.99.1]
 
 ### Changed